+++ tests/test_search.py
@@ -639,6 +639,66 @@ def test_search_result_with_exact_modifier(es_data, engine):
     assert bundle.total == 0
 
 
+def test_search_identifier_modifier(es_data, engine):
+    search_context = SearchContext(engine, "Observation")
+
+    params = (("subject:identifier", "240365-0002"),)
+    bundle = Search(search_context, params=params)()
+    assert bundle.total == 1
+
+    params = (("subject:identifier", "123465789"),)
+    bundle = Search(search_context, params=params)()
+    assert bundle.total == 0
+
+    # [param-ref]:identifier=[system]|[value]: the value of [code] matches an
+    # reference.identifier.value, and the value of [system] matches the system
+    # property of the Identifier
+    params = (
+        (
+            "subject:identifier",
+            "CPR|240365-0002",
+        ),
+    )
+    bundle = Search(search_context, params=params)()
+    assert bundle.total == 1
+
+    params = (
+        (
+            "subject:identifier",
+            "CPR|123456789",
+        ),
+    )
+    bundle = Search(search_context, params=params)()
+    assert bundle.total == 0
+
+    # TODO: not working yet, when omitting the system, the search is applied only on the value
+    # (it should also filter by empty system)
+    # [param-ref]:identifier=|[code]: the value of [code] matches a reference.identifier.value,
+    # and the Identifier has no system property
+    params = (("subject:identifier", "|240365-0002"),)
+    bundle = Search(search_context, params=params)()
+    assert bundle.total == 1
+
+    params = (("subject:identifier", "|123456789"),)
+    bundle = Search(search_context, params=params)()
+    assert bundle.total == 0
+
+    # [param-ref]:identifier=[system]|: any element where the value of [system] matches the
+    # system property of the Identifier
+    params = (
+        (
+            "subject:identifier",
+            "CPR|",
+        ),
+    )
+    bundle = Search(search_context, params=params)()
+    assert bundle.total == 1
+
+    params = (("subject:identifier", "other|"),)
+    bundle = Search(search_context, params=params)()
+    assert bundle.total == 0
+
+
 def test_issue9_multiple_negative_terms_not_working(es_data, engine):
     """https://github.com/nazrulworld/fhirpath/issues/9"""
     search_context = SearchContext(engine, "Task")